+++ e2e/Cargo.toml
@@ -23,10 +23,6 @@ fuels = { workspace = true }
 tai64 = { workspace = true }
 tempfile = { workspace = true }
 tokio = { workspace = true, features = ["test-util"] }
-# TODO: segfault remove
-fuel-abi-types = { workspace = true }
-# TODO: segfault remove
-itertools = { workspace = true, features = ["use_std"] }
 
 [build-dependencies]
 anyhow = { workspace = true, features = ["std"] }
