+++ packages/fuels-programs/src/contract/loader.rs
@@ -1,6 +1,5 @@
 use std::collections::HashSet;
 
-use fuel_asm::{op, Instruction, RegId};
 use fuel_tx::{Bytes32, ContractId, Salt, StorageSlot};
 use fuels_accounts::Account;
 use fuels_core::{
@@ -13,70 +12,9 @@ use fuels_core::{
     },
 };
 
+use crate::asm_scripts::loader_contract_asm;
+
 use super::{compute_contract_id_and_state_root, Contract, Regular};
-
-// Creates a contract that loads the specified blobs into memory and delegates the call to the code contained in the blobs.
-pub fn loader_contract_asm(blob_ids: &[BlobId]) -> Result<Vec<u8>> {
-    const BLOB_ID_SIZE: u16 = 32;
-    let get_instructions = |num_of_instructions, num_of_blobs| {
-        // There are 2 main steps:
-        // 1. Load the blob contents into memory
-        // 2. Jump to the beginning of the memory where the blobs were loaded
-        // After that the execution continues normally with the loaded contract reading our
-        // prepared fn selector and jumps to the selected contract method.
-        [
-            // 1. Load the blob contents into memory
-            // Find the start of the hardcoded blob IDs, which are located after the code ends.
-            op::move_(0x10, RegId::PC),
-            // 0x10 to hold the address of the current blob ID.
-            op::addi(0x10, 0x10, num_of_instructions * Instruction::SIZE as u16),
-            // The contract is going to be loaded from the current value of SP onwards, save
-            // the location into 0x16 so we can jump into it later on.
-            op::move_(0x16, RegId::SP),
-            // Loop counter.
-            op::movi(0x13, num_of_blobs),
-            // LOOP starts here.
-            // 0x11 to hold the size of the current blob.
-            op::bsiz(0x11, 0x10),
-            // Push the blob contents onto the stack.
-            op::ldc(0x10, 0, 0x11, 1),
-            // Move on to the next blob.
-            op::addi(0x10, 0x10, BLOB_ID_SIZE),
-            // Decrement the loop counter.
-            op::subi(0x13, 0x13, 1),
-            // Jump backwards (3+1) instructions if the counter has not reached 0.
-            op::jnzb(0x13, RegId::ZERO, 3),
-            // 2. Jump into the memory where the contract is loaded.
-            // What follows is called _jmp_mem by the sway compiler.
-            // Subtract the address contained in IS because jmp will add it back.
-            op::sub(0x16, 0x16, RegId::IS),
-            // jmp will multiply by 4, so we need to divide to cancel that out.
-            op::divi(0x16, 0x16, 4),
-            // Jump to the start of the contract we loaded.
-            op::jmp(0x16),
-        ]
-    };
-
-    let num_of_instructions = u16::try_from(get_instructions(0, 0).len())
-        .expect("to never have more than u16::MAX instructions");
-
-    let num_of_blobs = u32::try_from(blob_ids.len()).map_err(|_| {
-        error!(
-            Other,
-            "the number of blobs ({}) exceeds the maximum number of blobs supported: {}",
-            blob_ids.len(),
-            u32::MAX
-        )
-    })?;
-
-    let instruction_bytes = get_instructions(num_of_instructions, num_of_blobs)
-        .into_iter()
-        .flat_map(|instruction| instruction.to_bytes());
-
-    let blob_bytes = blob_ids.iter().flatten().copied();
-
-    Ok(instruction_bytes.chain(blob_bytes).collect())
-}
 
 #[derive(Debug, Clone)]
 pub struct BlobsUploaded {
