+++ wasm-tests/Cargo.toml
@@ -13,6 +13,7 @@ rust-version = { workspace = true }
 crate-type = ['cdylib']
 
 [dev-dependencies]
+hex = { workspace = true }
 fuels = { workspace = true }
 fuels-core = { workspace = true }
 getrandom = { version = "0.2.11", features = ["js"] }
