+++ wasm-tests/src/lib.rs
@@ -8,8 +8,10 @@ mod tests {
         accounts::predicate::Predicate,
         core::{codec::ABIEncoder, traits::Tokenizable},
         macros::wasm_abigen,
-        types::{bech32::Bech32Address, errors::Result},
+        programs::debug::{parse_script, ScriptType},
+        types::{bech32::Bech32Address, errors::Result, AssetId, ContractId},
     };
+    use fuels_core::codec::runtime_decoder::{self, RuntimeDecoder};
     use wasm_bindgen_test::wasm_bindgen_test;
 
     #[wasm_bindgen_test]
@@ -209,6 +211,240 @@ mod tests {
         )?;
 
         assert_eq!(*predicate.address(), expected_address);
+
+        Ok(())
+    }
+
+    #[wasm_bindgen_test]
+    fn can_decode_a_contract_calling_script() -> Result<()> {
+        let script = hex::decode("724028d8724428b05d451000724828b82d41148a724029537244292b5d451000724829332d41148a24040000")?;
+        let script_data = hex::decode("000000000000000a00000000000000000000000000000000000000000000000000000000000000001e62ecaa5c32f1e51954f46149d5e542472bdba45838199406464af46ab147ed000000000000290800000000000029260000000000000016636865636b5f7374727563745f696e746567726974790000000201000000000000001400000000000000000000000000000000000000000000000000000000000000001e62ecaa5c32f1e51954f46149d5e542472bdba45838199406464af46ab147ed000000000000298300000000000029a20000000000000017695f616d5f63616c6c65645f646966666572656e746c7900000002011e62ecaa5c32f1e51954f46149d5e542472bdba45838199406464af46ab147ed000000000000007b00000000000001c8")?;
+
+        let abi = r#"{
+            "programType": "contract",
+            "specVersion": "1",
+            "encodingVersion": "1",
+            "concreteTypes": [
+                {
+                "type": "()",
+                "concreteTypeId": "2e38e77b22c314a449e91fafed92a43826ac6aa403ae6a8acb6cf58239fbaf5d"
+                },
+                {
+                "type": "bool",
+                "concreteTypeId": "b760f44fa5965c2474a3b471467a22c43185152129295af588b022ae50b50903"
+                },
+                {
+                "type": "struct AllStruct",
+                "concreteTypeId": "91804f0112892169cddf041007c9f16f95281d45c3f363e544c33dffc8179266",
+                "metadataTypeId": 1
+                },
+                {
+                "type": "struct CallData",
+                "concreteTypeId": "c1b2644ef8de5c5b7a95aaadf3f5cedd40f42286d459bcd051c3cc35fa1ce5ec",
+                "metadataTypeId": 2
+                },
+                {
+                "type": "struct MemoryAddress",
+                "concreteTypeId": "0b7b6a791f80f65fe493c3e0d0283bf8206871180c9b696797ff0098ff63b474",
+                "metadataTypeId": 3
+                }
+            ],
+            "metadataTypes": [
+                {
+                "type": "b256",
+                "metadataTypeId": 0
+                },
+                {
+                "type": "struct AllStruct",
+                "metadataTypeId": 1,
+                "components": [
+                    {
+                    "name": "some_struct",
+                    "typeId": 4
+                    }
+                ]
+                },
+                {
+                "type": "struct CallData",
+                "metadataTypeId": 2,
+                "components": [
+                    {
+                    "name": "memory_address",
+                    "typeId": 3
+                    },
+                    {
+                    "name": "num_coins_to_forward",
+                    "typeId": 7
+                    },
+                    {
+                    "name": "asset_id_of_coins_to_forward",
+                    "typeId": 5
+                    },
+                    {
+                    "name": "amount_of_gas_to_forward",
+                    "typeId": 7
+                    }
+                ]
+                },
+                {
+                "type": "struct MemoryAddress",
+                "metadataTypeId": 3,
+                "components": [
+                    {
+                    "name": "contract_id",
+                    "typeId": 5
+                    },
+                    {
+                    "name": "function_selector",
+                    "typeId": 7
+                    },
+                    {
+                    "name": "function_data",
+                    "typeId": 7
+                    }
+                ]
+                },
+                {
+                "type": "struct SomeStruct",
+                "metadataTypeId": 4,
+                "components": [
+                    {
+                    "name": "field",
+                    "typeId": 6
+                    },
+                    {
+                    "name": "field_2",
+                    "typeId": "b760f44fa5965c2474a3b471467a22c43185152129295af588b022ae50b50903"
+                    }
+                ]
+                },
+                {
+                "type": "struct std::contract_id::ContractId",
+                "metadataTypeId": 5,
+                "components": [
+                    {
+                    "name": "bits",
+                    "typeId": 0
+                    }
+                ]
+                },
+                {
+                "type": "u32",
+                "metadataTypeId": 6
+                },
+                {
+                "type": "u64",
+                "metadataTypeId": 7
+                }
+            ],
+            "functions": [
+                {
+                "inputs": [
+                    {
+                    "name": "arg",
+                    "concreteTypeId": "91804f0112892169cddf041007c9f16f95281d45c3f363e544c33dffc8179266"
+                    }
+                ],
+                "name": "check_struct_integrity",
+                "output": "b760f44fa5965c2474a3b471467a22c43185152129295af588b022ae50b50903",
+                "attributes": [
+                    {
+                    "name": "payable",
+                    "arguments": []
+                    }
+                ]
+                },
+                {
+                "inputs": [],
+                "name": "get_struct",
+                "output": "91804f0112892169cddf041007c9f16f95281d45c3f363e544c33dffc8179266",
+                "attributes": null
+                },
+                {
+                "inputs": [
+                    {
+                    "name": "arg1",
+                    "concreteTypeId": "91804f0112892169cddf041007c9f16f95281d45c3f363e544c33dffc8179266"
+                    },
+                    {
+                    "name": "arg2",
+                    "concreteTypeId": "0b7b6a791f80f65fe493c3e0d0283bf8206871180c9b696797ff0098ff63b474"
+                    }
+                ],
+                "name": "i_am_called_differently",
+                "output": "2e38e77b22c314a449e91fafed92a43826ac6aa403ae6a8acb6cf58239fbaf5d",
+                "attributes": [
+                    {
+                    "name": "payable",
+                    "arguments": []
+                    }
+                ]
+                },
+                {
+                "inputs": [
+                    {
+                    "name": "call_data",
+                    "concreteTypeId": "c1b2644ef8de5c5b7a95aaadf3f5cedd40f42286d459bcd051c3cc35fa1ce5ec"
+                    }
+                ],
+                "name": "nested_struct_with_reserved_keyword_substring",
+                "output": "c1b2644ef8de5c5b7a95aaadf3f5cedd40f42286d459bcd051c3cc35fa1ce5ec",
+                "attributes": null
+                }
+            ],
+            "loggedTypes": [],
+            "messagesTypes": [],
+            "configurables": []
+        }"#;
+
+        let decoder = RuntimeDecoder::from_json_abi(abi)?;
+
+        // when
+        let script_type = parse_script(&script, &script_data)?;
+
+        // then
+        let ScriptType::ContractCall(call_descriptions) = script_type else {
+            panic!("expected a contract call")
+        };
+
+        assert_eq!(call_descriptions.len(), 2);
+
+        let call_description = &call_descriptions[0];
+
+        let expected_contract_id =
+            "1e62ecaa5c32f1e51954f46149d5e542472bdba45838199406464af46ab147ed".parse()?;
+        assert_eq!(call_description.contract_id, expected_contract_id);
+        assert_eq!(call_description.amount, 10);
+        assert_eq!(call_description.asset_id, AssetId::default());
+        assert_eq!(
+            call_description.decode_fn_selector().unwrap(),
+            "check_struct_integrity"
+        );
+        assert!(call_description.gas_forwarded.is_none());
+
+        assert_eq!(
+            decoder.decode_fn_args(
+                &call_description.decode_fn_selector().unwrap(),
+                &call_description.encoded_args
+            )?,
+            vec!["AllStruct { some_struct: SomeStruct { field: 2, field_2: true } }"]
+        );
+
+        let call_description = &call_descriptions[1];
+
+        assert_eq!(call_description.contract_id, expected_contract_id);
+        assert_eq!(call_description.amount, 20);
+        assert_eq!(call_description.asset_id, AssetId::default());
+        assert_eq!(
+            call_description.decode_fn_selector().unwrap(),
+            "i_am_called_differently"
+        );
+        assert!(call_description.gas_forwarded.is_none());
+
+        assert_eq!(
+            decoder.decode_fn_args(&call_description.decode_fn_selector().unwrap(), &call_description.encoded_args)?,
+            vec!["AllStruct { some_struct: SomeStruct { field: 2, field_2: true } }", "MemoryAddress { contract_id: std::contract_id::ContractId { bits: Bits256([30, 98, 236, 170, 92, 50, 241, 229, 25, 84, 244, 97, 73, 213, 229, 66, 71, 43, 219, 164, 88, 56, 25, 148, 6, 70, 74, 244, 106, 177, 71, 237]) }, function_selector: 123, function_data: 456 }"]
+        );
 
         Ok(())
     }
